v1.0.3

+++ ScottNet.Web/ViewModels/WeatherReadingViewModel.cs
@@ -9,25 +9,25 @@ namespace ScottNet.Web.ViewModels
     {
         public DateTime? ConsoleTime { get; set; }
         public int BarometerTrendCode { get; set; }
-        public double Barometer { get; set; }
-        public double IndoorTemp { get; set; }
+        public decimal Barometer { get; set; }
+        public decimal IndoorTemp { get; set; }
         public int IndoorHumidity { get; set; }
         public int OutdoorHumidity { get; set; }
-        public double OutdoorTemp { get; set; }
+        public decimal OutdoorTemp { get; set; }
         public int WindDirection { get; set; }
         public int WindSpeed { get; set; }
-        public double WindSpeed10M { get; set; }
-        public double WindSpeed2M { get; set; }
+        public decimal WindSpeed10M { get; set; }
+        public decimal WindSpeed2M { get; set; }
         public int WindGustSpeed10M { get; set; }
         public int WindGustDirection10M { get; set; }
-        public double RainRate { get; set; }
-        public double RainStorm { get; set; }
-        public double RainDay { get; set; }
-        public double RainMonth { get; set; }
-        public double RainYear { get; set; }
-        public double Rain15M { get; set; }
-        public double Rain1H { get; set; }
-        public double Rain24H { get; set; }
+        public decimal RainRate { get; set; }
+        public decimal RainStorm { get; set; }
+        public decimal RainDay { get; set; }
+        public decimal RainMonth { get; set; }
+        public decimal RainYear { get; set; }
+        public decimal Rain15M { get; set; }
+        public decimal Rain1H { get; set; }
+        public decimal Rain24H { get; set; }
         public int DewPoint { get; set; }
         public int HeatIndex { get; set; }
         public int WindChill { get; set; }
